backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.95.0
uvicorn[standard]>=0.21.1

# Database
sqlalchemy>=2.0.9
psycopg2-binary>=2.9.6
alembic>=1.10.3

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Utilities
python-dateutil>=2.8.2
pydantic[email]>=1.10.7
requests>=2.28.2
pytz>=2023.3

# QR Code generation
qrcode>=7.4.2
Pillow>=9.5.0

# Environment variables
python-dotenv>=1.0.0

# CORS
cors>=1.0.1

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.0.0

# Development tools
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.2.0
pre-commit>=3.2.2
pydantic-settings
